- contents(atPath:) to use Range instead of offset-length pair - write(data:) parameter is DataProtocol - Updated Readme - Refactoring obj-c layer
